feat: implemented pagination (#44)

+++ migrations/20220405143703_create_topic_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS topic
 (
     id          serial                                             NOT NULL PRIMARY KEY,
     name        varchar(50)                                        NOT NULL,
-    language_id int                                                NOT NULL REFERENCES language (id),
+    language_id smallint                                           NOT NULL REFERENCES language (id),
     created_at  timestamptz DEFAULT current_timestamp NOT NULL,
     updated_at  timestamptz DEFAULT current_timestamp NOT NULL
 );
